+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: `Carousel` now passes props to its children.
+- Feature: `Carousel` passes props relating to visibility to its children.
 
 ## [2.77.0] - 2020-07-02